refactor: move agoric config to agoric folder
